Added main.py
Created a basic outline and plotting capabilities for displaying the output calculations. Initial generation script for values isn't great but it is displaying values properly. May need to add better range to make it more ideal for our initial testing or use fixed variables. Linear algebra for the calculations should be able to be done with the np system. Option 1 is do-able now, option 2 is more complex but will be using an energy conservative approach for using fins instead of assuming.